+++ diskette/core/loader.py
@@ -2,12 +2,16 @@ import json
 import shutil
 import tarfile
 import tempfile
+import requests
 from pathlib import Path
 
+from django.conf import settings
 from django.template.defaultfilters import filesizeformat
 
 from ..utils.filesystem import directory_size
 from ..utils.loggers import NoOperationLogger
+from ..utils import hashs
+from ..utils.http import is_url
 
 from .serializers import LoaddataSerializerAbstract
 from .storages import StorageMixin
@@ -25,11 +29,50 @@ class Loader(StorageMixin, LoaddataSerializerAbstract):
     """
     MANIFEST_FILENAME = "manifest.json"
     TEMPDIR_PREFIX = "diskette_"
+    DOWNLOAD_FILENAME = "diskette_downloaded_archive.tar.gz"
 
     def __init__(self, logger=None):
         self.logger = logger or NoOperationLogger()
 
-    def open(self, archive_path, keep=False):
+    def download_archive(self, url, destination=None):
+        """
+        Download archive from given URL into destination directory.
+
+        Arguments:
+            url (string): The archive URL to download.
+
+        Keyword Arguments:
+            destination (Path): A path where to write downloaded archive file. If not
+                given, the archive file will be written as
+                ``diskette_downloaded_archive.tar.gz`` into the current working
+                directory.
+
+        Returns:
+            Path: Path to downloaded archive file.
+        """
+        destination = destination or Path.cwd() / self.DOWNLOAD_FILENAME
+        self.logger.info(
+            "Downloading archive from '{}' to '{}'".format(url, destination)
+        )
+
+        request = requests.get(
+            url,
+            allow_redirects=settings.DISKETTE_DOWNLOAD_ALLOW_REDIRECT,
+            timeout=settings.DISKETTE_DOWNLOAD_TIMEOUT,
+            stream=True
+        )
+        request.raise_for_status()
+
+        # Use the chunk way to avoid retaining the whole file in memory
+        with open(destination, "wb") as fd:
+            for chunk in request.iter_content(
+                chunk_size=settings.DISKETTE_DOWNLOAD_CHUNK
+            ):
+                fd.write(chunk)
+
+        return destination
+
+    def open(self, source, download_destination=None, keep=False, checksum=None):
         """
         Extract archive files in a temporary directory.
 
@@ -40,30 +83,62 @@ class Loader(StorageMixin, LoaddataSerializerAbstract):
             directories.
 
         Arguments:
-            archive_path (Path): A Path object to the archive to open.
+            source (Path): A Path object to the archive to open.
+
+        Keyword Arguments:
+            download_destination (Path): A path where to write downloaded archive file.
+                If not given, the archive file will be written as
+                ``diskette_downloaded_archive.tar.gz`` into the current working
+                directory. This argument is useless with local archive file.
             keep (boolean): Archive won't be removed from filesystem if True, else the
                 archive file is removed once it have been extracted.
+            checksum (object): Manage if archive is checksumed or not depending value:
+
+                * If ``None``: Checksum is done and just output to logs;
+                * If ``True``: Checksum is done and just output to logs;
+                * If ``False``: No checksum are done or compared;
+                * Any other value is assumed to be a string for a checksum to compare.
+                  Then a checksum is done on archive and compared to the given one, if
+                  comparaison fails it results to a critical error.
 
         Returns:
             Path: The temporary directory where archive files have been extracted.
         """
-        if (
-            isinstance(archive_path, str) and
-            archive_path.startswith(("http://", "https://"))
-        ):
-            # TODO: Download file with request in a temporary path, to remove once
-            # extracted
-            # archive_path = Path(..)
-            raise NotImplementedError("Loader does not support archive URL yet.")
+        archive = source
+        if is_url(source):
+            archive = self.download_archive(source, destination=download_destination)
 
+        # The temporary directory where to extract archive content
         destination_tmpdir = Path(tempfile.mkdtemp(prefix=self.TEMPDIR_PREFIX))
 
-        # Extract everything in temporary directory
-        with tarfile.open(archive_path, "r:gz") as archive:
-            archive.extractall(destination_tmpdir)
+        # Perform checksum if not explicitely disabled
+        if checksum is not False:
+            archive_checksum = hashs.file_checksum(archive)
+            self.logger.debug(
+                "Archive checksum: {}".format(archive_checksum)
+            )
+            # Compare checksums if any
+            if checksum and checksum is not True:
+                if archive_checksum != checksum:
+                    self.logger.critical(
+                        "Checksums do not match. Your archive file is probably "
+                        "corrupted."
+                    )
 
-        if not keep:
-            archive_path.unlink()
+        try:
+            # Extract everything in temporary directory
+            with tarfile.open(archive, "r:gz") as archive_fp:
+                archive_fp.extractall(destination_tmpdir)
+        except Exception as e:
+            # Remove destination_tmpdir on extraction failure
+            if destination_tmpdir.exists():
+                shutil.rmtree(destination_tmpdir)
+            # Then raise the exception
+            raise e
+        finally:
+            # Remove archive if not required to be keeped
+            if not keep:
+                archive.unlink()
 
         return destination_tmpdir
 
@@ -210,26 +285,45 @@ class Loader(StorageMixin, LoaddataSerializerAbstract):
             for dump in manifest["datas"]
         ]
 
-    def deploy(self, archive_path, storages_destination, with_data=True,
-               with_storages=True, keep=False):
+    def deploy(self, archive, storages_destination, with_data=True, with_storages=True,
+               download_destination=None, keep=False, checksum=None):
         """
         Load archive and deploy its content.
 
         Arguments:
-            archive_path (Path): The tarball archive to open and extract dumps.
+            archive (Path or string): The tarball archive to open and extract dumps. It
+                may be either a Path to a local archive file or a string for an URL
+                to download the archive.
             storages_destination (Path): Destination where to deploy all storage
                 directories.
 
         Keyword Arguments:
             with_data (boolean): Enable application datas loading.
             with_storages (boolean): Enabled media storages loading.
+            download_destination (Path): A path where to write downloaded archive file.
+                If not given, the archive file will be written as
+                ``diskette_downloaded_archive.tar.gz`` into the current working
+                directory. This argument is useless with local archive file.
             keep (boolean): Archive won't be removed from filesystem if True, else the
                 archive file is removed once it have been extracted.
+            checksum (object): Manage if archive is checksumed or not depending value:
+
+                * If ``None``: Checksum is done and just output to logs;
+                * If ``True``: Checksum is done and just output to logs;
+                * If ``False``: No checksum are done or compared;
+                * Any other value is assumed to be a string for a checksum to compare.
+                  Then a checksum is done on archive and compared to the given one, if
+                  comparaison fails it results to a critical error.
 
         Returns:
             dict: Statistics of deployed storages and datas.
         """
-        tmpdir = self.open(archive_path, keep=keep)
+        tmpdir = self.open(
+            archive,
+            download_destination=download_destination,
+            keep=keep,
+            checksum=checksum,
+        )
 
         try:
             manifest = self.get_manifest(tmpdir)