+++ CHANGELOG.rst
@@ -3,6 +3,14 @@
 Changelog
 =========
 
+Version 0.3.2 - Unreleased
+--------------------------
+
+* Added support of archive URL to download in ``diskette_load``;
+* Added options for archive checksum creation and comparison in ``diskette_load``;
+* Added option for archive checksum creation in ``diskette_dump``;
+
+
 Version 0.3.1 - 2024/03/21
 --------------------------